+++ inst/examples/select_panel/app.R
@@ -77,7 +77,7 @@ server <- function(input, output, session) {
           )
         )
       ),
-      theme = "replit"
+      theme = "light-spaced"
     )
   })